chore: create file for tree hash function
