+++ crates/committer/src/patricia_merkle_tree/updated_skeleton_tree.rs
@@ -1,2 +1,3 @@
+pub mod hash_function;
 pub mod node;
 pub mod tree;
